Refactor to own plugin
